# openapi-generator/stub/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "DotAAS Part 2 | HTTP/REST | Asset Administration Shell Repository Service Specification",
    "description" : "The Full Profile of the Asset Administration Shell Repository Service Specification as part of the [Specification of the Asset Administration Shell: Part 2](http://industrialdigitaltwin.org/en/content-hub).  \nPublisher: Industrial Digital Twin Association (IDTA) April 2023",
    "contact" : {
      "name" : "Industrial Digital Twin Association (IDTA)",
      "email" : "info@idtwin.org"
    },
    "license" : {
      "name" : "CC BY 4.0",
      "url" : "https://creativecommons.org/licenses/by/4.0/"
    },
    "version" : "V3.0.1_SSP-001"
  },
  "servers" : [ {
    "url" : "{protocol}://{host_name}:{port}/api/{version_prefix}",
    "variables" : {
      "protocol" : {
        "description" : "Allows access through http and https (recommended)",
        "default" : "https",
        "enum" : [ "http", "https" ]
      },
      "host_name" : {
        "description" : "Hostname of server hosting the api",
        "default" : "admin-shell.io"
      },
      "port" : {
        "description" : "80 is default for http, 443 for https",
        "default" : "443",
        "enum" : [ "80", "443" ]
      },
      "version_prefix" : {
        "default" : "v3.0",
        "enum" : [ "v3.0" ]
      }
    }
  } ],
  "paths" : {
...
    "/serialization" : {
      "get" : {
        "tags" : [ "Serialization API" ],
        "summary" : "Returns an appropriate serialization based on the specified format (see SerializationFormat)",
        "operationId" : "GenerateSerializationByIds",
        "parameters" : [ {
          "name" : "aasIds",
          "in" : "query",
          "description" : "The Asset Administration Shells' unique ids (UTF8-BASE64-URL-encoded)",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "submodelIds",
          "in" : "query",
          "description" : "The Submodels' unique ids (UTF8-BASE64-URL-encoded)",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "includeConceptDescriptions",
          "in" : "query",
          "description" : "Include Concept Descriptions?",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Requested serialization based on SerializationFormat",
            "content" : {
              "application/asset-administration-shell-package+xml" : {
                "schema" : {
                  "type" : "string",
                  "description" : "AASX package",
                  "format" : "binary",
                  "x-content-type" : "application/asset-administration-shell-package+xml"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Environment"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Environment"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request, e.g. the request parameters of the format of the request body is wrong.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized, e.g. the server refused the authorization attempt.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "default" : {
            "description" : "Default error handling for unmentioned status codes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/GenerateSerializationByIds/3/0" ]
      }
    },
    "/description" : {
      "get" : {
        "tags" : [ "Description API" ],
        "summary" : "Returns the self-describing information of a network resource (ServiceDescription)",
        "operationId" : "GetDescription",
        "responses" : {
          "200" : {
            "description" : "Requested Description",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceDescription"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized, e.g. the server refused the authorization attempt.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/Descriptor/GetDescription/3/0" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "AssetAdministrationShell" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Identifiable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "required" : [ "assetInformation" ],
          "properties" : {
            "derivedFrom" : {
              "$ref" : "#/components/schemas/Reference"
            },
            "assetInformation" : {
              "$ref" : "#/components/schemas/AssetInformation"
            },
            "submodels" : {
              "minItems" : 1,
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Reference"
              }
            },
            "modelType" : {
              "pattern" : "AssetAdministrationShell",
              "type" : "string"
            }
          }
        } ]
      },
      "Identifiable" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Referable"
        }, {
          "required" : [ "id" ],
          "properties" : {
            "administration" : {
              "$ref" : "#/components/schemas/AdministrativeInformation"
            },
            "id" : {
              "maxLength" : 2000,
              "minLength" : 1,
              "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            }
          }
        } ]
      },
      "Referable" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/HasExtensions"
        }, {
          "required" : [ "modelType" ],
          "properties" : {
            "category" : {
              "maxLength" : 128,
              "minLength" : 1,
              "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            },
            "idShort" : {
              "type" : "string",
              "allOf" : [ {
                "maxLength" : 128,
                "minLength" : 1
              }, {
                "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$"
              }, {
                "pattern" : "^[a-zA-Z][a-zA-Z0-9_]*$"
              } ]
            },
            "displayName" : {
              "minItems" : 1,
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangStringNameType"
              }
            },
            "description" : {
              "minItems" : 1,
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangStringTextType"
              }
            },
            "modelType" : {
              "$ref" : "#/components/schemas/ModelType"
            }
          }
        } ]
      },
...
      "HasSemantics" : {
        "type" : "object",
        "properties" : {
          "semanticId" : {
            "$ref" : "#/components/schemas/Reference"
          },
          "supplementalSemanticIds" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          }
        }
      },
      "Reference" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/ReferenceParent"
        }, {
          "properties" : {
            "referredSemanticId" : {
              "$ref" : "#/components/schemas/ReferenceParent"
            }
          }
        } ]
      },
      "ReferenceParent" : {
        "required" : [ "keys", "type" ],
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/ReferenceTypes"
          },
          "keys" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Key"
            }
          }
        }
      },
...
      "ModelType" : {
        "type" : "string",
        "enum" : [ "AnnotatedRelationshipElement", "AssetAdministrationShell", "BasicEventElement", "Blob", "Capability", "ConceptDescription", "DataSpecificationIec61360", "Entity", "File", "MultiLanguageProperty", "Operation", "Property", "Range", "ReferenceElement", "RelationshipElement", "Submodel", "SubmodelElementCollection", "SubmodelElementList" ]
      },
      "AdministrativeInformation" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "properties" : {
            "version" : {
              "type" : "string",
              "allOf" : [ {
                "maxLength" : 4,
                "minLength" : 1
              }, {
                "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$"
              }, {
                "pattern" : "^(0|[1-9][0-9]*)$"
              } ]
            },
            "revision" : {
              "type" : "string",
              "allOf" : [ {
                "maxLength" : 4,
                "minLength" : 1
              }, {
                "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$"
              }, {
                "pattern" : "^(0|[1-9][0-9]*)$"
              } ]
            },
            "creator" : {
              "$ref" : "#/components/schemas/Reference"
            },
            "templateId" : {
              "maxLength" : 2000,
              "minLength" : 1,
              "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            }
          }
        } ]
      },
      "HasDataSpecification" : {
        "type" : "object",
        "properties" : {
          "embeddedDataSpecifications" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EmbeddedDataSpecification"
            }
          }
        }
      },
      "EmbeddedDataSpecification" : {
        "required" : [ "dataSpecification", "dataSpecificationContent" ],
        "type" : "object",
        "properties" : {
          "dataSpecification" : {
            "$ref" : "#/components/schemas/Reference"
          },
          "dataSpecificationContent" : {
            "$ref" : "#/components/schemas/DataSpecificationContent_choice"
          }
        }
      },
      "DataSpecificationContent_choice" : {
        "oneOf" : [ {
          "$ref" : "#/components/schemas/DataSpecificationIec61360"
        } ]
      },
...
      "AssetInformation" : {
        "required" : [ "assetKind" ],
        "type" : "object",
        "properties" : {
          "assetKind" : {
            "$ref" : "#/components/schemas/AssetKind"
          },
          "globalAssetId" : {
            "maxLength" : 2000,
            "minLength" : 1,
            "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
            "type" : "string"
          },
          "specificAssetIds" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SpecificAssetId"
            }
          },
          "assetType" : {
            "maxLength" : 2000,
            "minLength" : 1,
            "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
            "type" : "string"
          },
          "defaultThumbnail" : {
            "$ref" : "#/components/schemas/Resource"
          }
        },
        "example" : {
          "assetKind" : "Instance",
          "specificAssetIds" : [ "", "" ],
          "defaultThumbnail" : {
            "path" : "",
            "contentType" : ""
          },
          "globalAssetId" : "globalAssetId",
          "assetType" : "assetType"
        }
      },
      "AssetKind" : {
        "type" : "string",
        "enum" : [ "Instance", "NotApplicable", "Type" ]
      },
      "SpecificAssetId" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/HasSemantics"
        }, {
          "required" : [ "name", "value" ],
          "properties" : {
            "name" : {
              "maxLength" : 64,
              "minLength" : 1,
              "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            },
            "value" : {
              "maxLength" : 2000,
              "minLength" : 1,
              "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            },
            "externalSubjectId" : {
              "$ref" : "#/components/schemas/Reference"
            }
          }
        } ]
      },
      "Resource" : {
        "required" : [ "path" ],
        "type" : "object",
        "properties" : {
          "path" : {
            "type" : "string",
            "allOf" : [ {
              "maxLength" : 2000,
              "minLength" : 1
            }, {
              "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$"
            }, {
              "pattern" : "^file:(//((localhost|(\\[((([0-9A-Fa-f]{1,4}:){6}([0-9A-Fa-f]{1,4}:[0-9A-Fa-f]{1,4}|([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]))|::([0-9A-Fa-f]{1,4}:){5}([0-9A-Fa-f]{1,4}:[0-9A-Fa-f]{1,4}|([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]))|([0-9A-Fa-f]{1,4})?::([0-9A-Fa-f]{1,4}:){4}([0-9A-Fa-f]{1,4}:[0-9A-Fa-f]{1,4}|([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]))|(([0-9A-Fa-f]{1,4}:)?[0-9A-Fa-f]{1,4})?::([0-9A-Fa-f]{1,4}:){3}([0-9A-Fa-f]{1,4}:[0-9A-Fa-f]{1,4}|([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]))|(([0-9A-Fa-f]{1,4}:){2}[0-9A-Fa-f]{1,4})?::([0-9A-Fa-f]{1,4}:){2}([0-9A-Fa-f]{1,4}:[0-9A-Fa-f]{1,4}|([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]))|(([0-9A-Fa-f]{1,4}:){3}[0-9A-Fa-f]{1,4})?::[0-9A-Fa-f]{1,4}:([0-9A-Fa-f]{1,4}:[0-9A-Fa-f]{1,4}|([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]))|(([0-9A-Fa-f]{1,4}:){4}[0-9A-Fa-f]{1,4})?::([0-9A-Fa-f]{1,4}:[0-9A-Fa-f]{1,4}|([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]))|(([0-9A-Fa-f]{1,4}:){5}[0-9A-Fa-f]{1,4})?::[0-9A-Fa-f]{1,4}|(([0-9A-Fa-f]{1,4}:){6}[0-9A-Fa-f]{1,4})?::)|[vV][0-9A-Fa-f]+\\.([a-zA-Z0-9\\-._~]|[!$&'()*+,;=]|:)+)\\]|([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])|([a-zA-Z0-9\\-._~]|%[0-9A-Fa-f][0-9A-Fa-f]|[!$&'()*+,;=])*)))?/((([a-zA-Z0-9\\-._~]|%[0-9A-Fa-f][0-9A-Fa-f]|[!$&'()*+,;=]|[:@]))+(/(([a-zA-Z0-9\\-._~]|%[0-9A-Fa-f][0-9A-Fa-f]|[!$&'()*+,;=]|[:@]))*)*)?|/((([a-zA-Z0-9\\-._~]|%[0-9A-Fa-f][0-9A-Fa-f]|[!$&'()*+,;=]|[:@]))+(/(([a-zA-Z0-9\\-._~]|%[0-9A-Fa-f][0-9A-Fa-f]|[!$&'()*+,;=]|[:@]))*)*)?)$"
            } ]
          },
          "contentType" : {
            "type" : "string",
            "allOf" : [ {
              "maxLength" : 100,
              "minLength" : 1
            }, {
              "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$"
            }, {
              "pattern" : "^([!#$%&'*+\\-.^_`|~0-9a-zA-Z])+/([!#$%&'*+\\-.^_`|~0-9a-zA-Z])+([ \\t]*;[ \\t]*([!#$%&'*+\\-.^_`|~0-9a-zA-Z])+=(([!#$%&'*+\\-.^_`|~0-9a-zA-Z])+|\"(([\\t !#-\\[\\]-~]|[-ÿ])|\\\\([\\t !-~]|[-ÿ]))*\"))*$"
            } ]
          }
        },
        "example" : {
          "path" : "",
          "contentType" : ""
        }
      },
      "Result" : {
        "type" : "object",
        "properties" : {
          "messages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Message"
            }
          }
        }
      },
      "Message" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "maxLength" : 32,
            "minLength" : 1,
            "type" : "string"
          },
          "correlationId" : {
            "maxLength" : 128,
            "minLength" : 1,
            "type" : "string"
          },
          "messageType" : {
            "type" : "string",
            "enum" : [ "Undefined", "Info", "Warning", "Error", "Exception" ]
          },
          "text" : {
            "type" : "string"
          },
          "timestamp" : {
            "pattern" : "^-?(([1-9][0-9][0-9][0-9]+)|(0[0-9][0-9][0-9]))-((0[1-9])|(1[0-2]))-((0[1-9])|([12][0-9])|(3[01]))T(((([01][0-9])|(2[0-3])):[0-5][0-9]:([0-5][0-9])(\\.[0-9]+)?)|24:00:00(\\.0+)?)(Z|\\+00:00|-00:00)$",
            "type" : "string"
          }
        }
      },
...
      "Submodel" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Identifiable"
        }, {
          "$ref" : "#/components/schemas/HasKind"
        }, {
          "$ref" : "#/components/schemas/HasSemantics"
        }, {
          "$ref" : "#/components/schemas/Qualifiable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "properties" : {
            "submodelElements" : {
              "minItems" : 1,
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/SubmodelElement_choice"
              }
            },
            "modelType" : {
              "pattern" : "Submodel",
              "type" : "string"
            }
          }
        } ]
      },
      "HasKind" : {
        "type" : "object",
        "properties" : {
          "kind" : {
            "$ref" : "#/components/schemas/ModellingKind"
          }
        }
      },
      "ModellingKind" : {
        "type" : "string",
        "enum" : [ "Instance", "Template" ]
      },
      "Qualifiable" : {
        "required" : [ "modelType" ],
        "type" : "object",
        "properties" : {
          "qualifiers" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Qualifier"
            }
          },
          "modelType" : {
            "$ref" : "#/components/schemas/ModelType"
          }
        }
      },
      "Qualifier" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/HasSemantics"
        }, {
          "required" : [ "type", "valueType" ],
          "properties" : {
            "kind" : {
              "$ref" : "#/components/schemas/QualifierKind"
            },
            "type" : {
              "maxLength" : 128,
              "minLength" : 1,
              "pattern" : "^([\\t\\n\\r -퟿-�]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            },
            "valueType" : {
              "$ref" : "#/components/schemas/DataTypeDefXsd"
            },
            "value" : {
              "type" : "string"
            },
            "valueId" : {
              "$ref" : "#/components/schemas/Reference"
            }
          }
        } ]
      },
...
      "SubmodelElement_choice" : {
        "oneOf" : [ {
          "$ref" : "#/components/schemas/RelationshipElement"
        }, {
          "$ref" : "#/components/schemas/AnnotatedRelationshipElement"
        }, {
          "$ref" : "#/components/schemas/BasicEventElement"
        }, {
          "$ref" : "#/components/schemas/Blob"
        }, {
          "$ref" : "#/components/schemas/Capability"
        }, {
          "$ref" : "#/components/schemas/Entity"
        }, {
          "$ref" : "#/components/schemas/File"
        }, {
          "$ref" : "#/components/schemas/MultiLanguageProperty"
        }, {
          "$ref" : "#/components/schemas/Operation"
        }, {
          "$ref" : "#/components/schemas/Property"
        }, {
          "$ref" : "#/components/schemas/Range"
        }, {
          "$ref" : "#/components/schemas/ReferenceElement"
        }, {
          "$ref" : "#/components/schemas/SubmodelElementCollection"
        }, {
          "$ref" : "#/components/schemas/SubmodelElementList"
        } ]
      },
      "RelationshipElement" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/RelationshipElement_abstract"
        }, {
          "properties" : {
            "modelType" : {
              "pattern" : "RelationshipElement",
              "type" : "string"
            }
          }
        } ]
      },
...
      "Environment" : {
        "type" : "object",
        "properties" : {
          "assetAdministrationShells" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AssetAdministrationShell"
            }
          },
          "submodels" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Submodel"
            }
          },
          "conceptDescriptions" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConceptDescription"
            }
          }
        }
      },
      "ConceptDescription" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Identifiable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "properties" : {
            "isCaseOf" : {
              "minItems" : 1,
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Reference"
              }
            },
            "modelType" : {
              "pattern" : "ConceptDescription",
              "type" : "string"
            }
          }
        } ]
      },
      "ServiceDescription" : {
        "type" : "object",
        "properties" : {
          "profiles" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "https://admin-shell.io/aas/API/3/0/AssetAdministrationShellServiceSpecification/SSP-001", "https://admin-shell.io/aas/API/3/0/AssetAdministrationShellServiceSpecification/SSP-002", "https://admin-shell.io/aas/API/3/0/SubmodelServiceSpecification/SSP-001", "https://admin-shell.io/aas/API/3/0/SubmodelServiceSpecification/SSP-002", "https://admin-shell.io/aas/API/3/0/SubmodelServiceSpecification/SSP-003", "https://admin-shell.io/aas/API/3/0/AasxFileServerServiceSpecification/SSP-001", "https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRegistryServiceSpecification/SSP-001", "https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRegistryServiceSpecification/SSP-002", "https://admin-shell.io/aas/API/3/0/SubmodelRegistryServiceSpecification/SSP-001", "https://admin-shell.io/aas/API/3/0/SubmodelRegistryServiceSpecification/SSP-002", "https://admin-shell.io/aas/API/3/0/DiscoveryServiceSpecification/SSP-001", "https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRepositoryServiceSpecification/SSP-001", "https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRepositoryServiceSpecification/SSP-002", "https://admin-shell.io/aas/API/3/0/SubmodelRepositoryServiceSpecification/SSP-001", "https://admin-shell.io/aas/API/3/0/SubmodelRepositoryServiceSpecification/SSP-002", "https://admin-shell.io/aas/API/3/0/SubmodelRepositoryServiceSpecification/SSP-003", "https://admin-shell.io/aas/API/3/0/SubmodelRepositoryServiceSpecification/SSP-004", "https://admin-shell.io/aas/API/3/0/ConceptDescriptionServiceSpecification/SSP-001" ]
            }
          }
        },
        "description" : "The Description object enables servers to present their capabilities to the clients, in particular which profiles they implement. At least one defined profile is required. Additional, proprietary attributes might be included. Nevertheless, the server must not expect that a regular client understands them.",
        "example" : "{\n  \"profiles\": [\n    \"https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRegistryServiceSpecification/SSP-002\",\n    \"https://admin-shell.io/aas/API/3/0/SubmodelRegistryServiceSpecification/SSP-002\"\n  ]\n}"
      },
...
    },
...
  }
}